rename components per Vue style guide
