+++ training-front-end/src/components/USWDSAlert.vue
