+++ training-front-end/src/components/icons/CancelIcon.vue
