+++ training-front-end/src/components/icons/PrintIcon.vue
